Refactor to create tree modules
