+++ cfgparser/nokia/classic/lexer.py
@@ -1,49 +1,11 @@
 from __future__ import annotations
 
 import typing as t
-from abc import ABC
-from abc import abstractmethod
+
+from cfgparser.tree.token import AbstractTokenBuilder
+from cfgparser.tree.token import Token
 
 INDENT_SZ = 4
-
-
-class Token:
-    def __init__(
-        self,
-        name: str,
-        value: t.Optional[str],
-        indent: int,
-        params: t.Optional[t.List] = None,
-        childs: t.Optional[t.Dict] = None,
-    ) -> None:
-        self.name: str = name
-        self.value: t.Optional[str] = value
-        self.indent: int = indent
-        self.params: list = []
-        self.childs: dict = {}
-        self.is_container: bool = False
-
-        if params:
-            self.params = params
-
-        if childs:
-            self.childs = childs
-
-    @property
-    def id(self) -> str:
-        if isinstance(self.value, str):
-            return f"{self.name} {self.value}"
-        return self.name
-
-
-class AbstractTokenBuilder(ABC):
-    @staticmethod
-    @abstractmethod
-    def check_rule(words: t.List[str]) -> bool: ...
-
-    @staticmethod
-    @abstractmethod
-    def create(words: t.List[str], indent: int) -> Token: ...
 
 
 class DefaultTokenBuilder(AbstractTokenBuilder):
@@ -240,7 +202,7 @@ class EpipeTokenBuilder(AbstractTokenBuilder):
     @staticmethod
     def create(words: list, indent: int) -> Token:
         name, value, __, epipe_name, __, cust_id, __ = words
-        
+
         token = Token(name, value, indent)
         token.is_container = True
